extract the interfaces back into the repository package
