+++ app/src/main/java/net/squanchy/injection/ApplicationComponent.kt
@@ -11,9 +11,9 @@ import net.squanchy.schedule.tracksfilter.TracksFilterModule
 import net.squanchy.service.firestore.FirebaseAuthService
 import net.squanchy.service.firestore.FirestoreDbService
 import net.squanchy.service.firestore.injection.FirestoreModule
-import net.squanchy.service.repository.firestore.EventRepository
-import net.squanchy.service.repository.firestore.SpeakerRepository
-import net.squanchy.service.repository.firestore.TracksRepository
+import net.squanchy.service.repository.EventRepository
+import net.squanchy.service.repository.SpeakerRepository
+import net.squanchy.service.repository.TracksRepository
 import net.squanchy.service.repository.injection.RepositoryModule
 import net.squanchy.support.injection.ChecksumModule
 import net.squanchy.support.injection.CurrentTimeModule
