+++ app/src/main/java/net/squanchy/notification/NotificationModule.kt
@@ -6,7 +6,7 @@ import dagger.Module
 import dagger.Provides
 import net.squanchy.injection.ServiceContextModule
 import net.squanchy.service.firestore.FirebaseAuthService
-import net.squanchy.service.repository.firestore.EventRepository
+import net.squanchy.service.repository.EventRepository
 
 @Module(includes = [ServiceContextModule::class])
 internal class NotificationModule {
