+++ app/src/main/java/net/squanchy/notification/NotificationService.kt
@@ -4,7 +4,7 @@ import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
 import net.squanchy.schedule.domain.view.Event
 import net.squanchy.service.firestore.FirebaseAuthService
-import net.squanchy.service.repository.firestore.EventRepository
+import net.squanchy.service.repository.EventRepository
 
 internal class NotificationService(private val authService: FirebaseAuthService, private val eventRepository: EventRepository) {
