+++ app/src/main/java/net/squanchy/service/repository/injection/RepositoryModule.kt
@@ -3,12 +3,12 @@ package net.squanchy.service.repository.injection
 import dagger.Module
 import dagger.Provides
 import net.squanchy.service.firestore.FirestoreDbService
-import net.squanchy.service.repository.firestore.EventRepository
+import net.squanchy.service.repository.EventRepository
 import net.squanchy.service.repository.firestore.FirestoreEventRepository
 import net.squanchy.service.repository.firestore.FirestoreSpeakerRepository
 import net.squanchy.service.repository.firestore.FirestoreTracksRepository
-import net.squanchy.service.repository.firestore.SpeakerRepository
-import net.squanchy.service.repository.firestore.TracksRepository
+import net.squanchy.service.repository.SpeakerRepository
+import net.squanchy.service.repository.TracksRepository
 import net.squanchy.support.lang.Checksum
 
 @Module
