+++ app/src/main/java/net/squanchy/speaker/SpeakerDetailsService.kt
@@ -2,7 +2,7 @@ package net.squanchy.speaker
 
 import io.reactivex.Observable
 import net.squanchy.service.firestore.FirebaseAuthService
-import net.squanchy.service.repository.firestore.SpeakerRepository
+import net.squanchy.service.repository.SpeakerRepository
 import net.squanchy.speaker.domain.view.Speaker
 
 internal class SpeakerDetailsService(
